Validations done

+++ src/main/java/com/hexaware/carrentalsystems/service/ReservationServiceImp.java
@@ -5,20 +5,47 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.hexaware.carrentalsystems.dto.ReservationDto;
+import com.hexaware.carrentalsystems.entities.Car;
 import com.hexaware.carrentalsystems.entities.Reservation;
+import com.hexaware.carrentalsystems.entities.User;
+import com.hexaware.carrentalsystems.exceptions.CarNotFoundException;
+import com.hexaware.carrentalsystems.exceptions.FeedbackNotFoundException;
 import com.hexaware.carrentalsystems.exceptions.ReservationNotFoundException;
+import com.hexaware.carrentalsystems.exceptions.UserNotFoundException;
+import com.hexaware.carrentalsystems.repository.ICarRepository;
 import com.hexaware.carrentalsystems.repository.IReservationRepository;
+import com.hexaware.carrentalsystems.repository.IUserRepository;
 
 @Service
 public class ReservationServiceImp implements IReservationService {
 
     @Autowired
     private IReservationRepository repo;
+    
+    @Autowired
+    IUserRepository UserRepo;
+
+    @Autowired
+    ICarRepository CarRepo;
 
     @Override
-    public Reservation addReservation(Reservation reservation) {
-        return repo.save(reservation);
+    public Reservation addReservation(ReservationDto dto) {
+        User user = UserRepo.findById(dto.getUserId()).orElseThrow(() -> new UserNotFoundException("User not found"));
+        Car car = CarRepo.findById(dto.getCarId()).orElseThrow(() -> new CarNotFoundException("Car not found"));
+
+        Reservation res = new Reservation();
+        res.setReservationId(dto.getReservationId());
+        res.setUser(user);
+        res.setCar(car);
+        res.setPickupDate(dto.getPickupDate());
+        res.setDropoffDate(dto.getDropoffDate());
+        res.setTotalAmount(dto.getTotalAmount());
+        res.setStatus(dto.getStatus());
+
+        return repo.save(res);
     }
+
 
     @Override
     public Reservation updateReservation(Reservation reservation) {
@@ -38,7 +65,12 @@ public class ReservationServiceImp implements IReservationService {
 
     @Override
     public String deleteByReservationId(int reservationId) {
-        repo.deleteById(reservationId);
-        return "Reservation deleted successfully";
+    	if (repo.existsById(reservationId)) {
+    	 repo.deleteById(reservationId);
+    	 return "Reservation deleted";
+    	 } else {
+    	        throw  new ReservationNotFoundException("Reservation not found with ID: " + reservationId);
     }
+    }  
 }
+
